Nominate Felix Palmer to the Technical Steering Committee
I have really enjoyed the opportunity to work with deck.gl and with such a talented group of people. I have learned on the technical side and how to define the strategy between a commercial company and an Open Source project. We at CARTO decided to switch our visualization strategy in favor of deck.gl and it has been a complete success for all the parts involved.

I can't continue as member of the technical steering committee because I'm starting a new project and won't have the time to dedicate to this project.

Felix Palmer is my natural successor and I have great confidence in him to perform carry out the tasks.

A big fan of vis.gl! Vamos!!